+++ src/commands/new-post.ts
@@ -4,10 +4,19 @@ import path from 'path';
 
 import {getHugoWorkspaceFolder, hugo} from '../utils';
 
-const getOffsetDateString = (): string => {
+type DateFormat = 'yyyymmdd'|'yyyy-mm-dd';
+
+const getOffsetDateString = (dateFormat: DateFormat|string): string => {
     const date = new Date();
     date.setMinutes(date.getMinutes() - date.getTimezoneOffset());
-    return date.toISOString().slice(0, 10).replace(/-/g, '');
+    const dateString = date.toISOString().slice(0, 10);
+    switch (dateFormat) {
+        case 'yyyymmdd':
+            return dateString.replace(/-/g, '');
+        case 'yyyy-mm-dd':
+            return dateString;
+    };
+    return dateString;
 };
 
 const newPost = async () => {
@@ -17,15 +26,16 @@ const newPost = async () => {
         return;
     }
 
-    vscode.commands.executeCommand('setContext', 'hugo-utils.inHugoWorkspace', true);
-
     // Prompt for the new post path
-    const date = getOffsetDateString();
     const section = vscode.workspace.getConfiguration('hugo-utils').get<string>('mainSectionName')!;
+
+    const dateFormat = vscode.workspace.getConfiguration('hugo-utils').get<DateFormat|string>('dateFormat')!;
+    const date = getOffsetDateString(dateFormat);
+
     const postName = await vscode.window.showInputBox({
         prompt: 'Path to the new post',
         value: `${section}/${date}-title.md`,
-        valueSelection: [section.length + 10, section.length + 15],
+        valueSelection: [section.length + date.length + 2, section.length + date.length + 2 + 5],
     });
     if (!postName) {
         return;